evaluate_model.py changes:
1. Remove unicode LTR/RTL before text evaluation.
2. Restructure code to invoke separate engines as callables.
